requirements.txt
# General dependencies for development environment.
# The add-on itself installs platform-specific versions via src/cgt_mediapipe/cgt_dependencies.py

absl-py
attrs>=19.1.0
matplotlib
opencv-contrib-python>=4.5.5.64
protobuf==3.20.2
mediapipe>=0.10.0
numpy>=1.24.0
